First past at refactoring to use context in Field
