+++ package.json
@@ -16,6 +16,7 @@
     "@radix-ui/react-accordion": "^1.0.1",
     "@radix-ui/react-label": "^2.0.0",
     "@radix-ui/react-select": "^1.1.1",
+    "@radix-ui/react-slot": "^1.0.1",
     "classnames": "^2.3.1",
     "date-fns": "^2.29.3",
     "eslint-config-next": "^13.2.2",
